this is so annoing

+++ DevDialogueAddon/Assets/DyConv/Scripts/DyConvWindow.cs
@@ -16,6 +16,8 @@ public class DyConvWindow : EditorWindow
     bool isActiveViewer = false;
     string currentDialogueName;
     int currentNextDialogueIndex = -1;
+    int currentSelectCriteriaDialogueIndex = -1;
+    int currentSelectCriteriaIndex = -1;
     List<bool> shownNextDialogueList = new List<bool>();
     List<bool> shownCriteriaList = new List<bool>();
     Vector2 criteriaScrollPos = new Vector2();
@@ -40,8 +42,11 @@ public class DyConvWindow : EditorWindow
 
     private void OnGUI()
     {
+
         EditorGUILayout.BeginHorizontal();
         EditorGUILayout.BeginVertical();
+
+        //Fact 이름과 값을 표시해주는 창
         #region Fact
         EditorGUILayout.BeginVertical(GUI.skin.box, GUILayout.MaxHeight(Screen.height / 2));
 
@@ -51,12 +56,13 @@ public class DyConvWindow : EditorWindow
 
         if (GUILayout.Button(EditorGUIUtility.IconContent("d_Toolbar Plus@2x").image, EditorStyles.iconButton))
         {
-            AddNewFactWindow.Open();
+            if(!Application.isPlaying)
+                AddNewFactWindow.Open();
         }
         EditorGUI.BeginDisabledGroup(string.IsNullOrEmpty(currentFactName));
         if (GUILayout.Button(EditorGUIUtility.IconContent("d_Toolbar Minus@2x").image, EditorStyles.iconButton))
         {
-            if (!string.IsNullOrEmpty(currentFactName))
+            if (!string.IsNullOrEmpty(currentFactName) && !Application.isPlaying)
             {
                 FactContainer.FactDictionary.Remove(currentFactName);
                 EditorUtility.SetDirty(FactContainer.container);
@@ -68,6 +74,7 @@ public class DyConvWindow : EditorWindow
         EditorGUI.EndDisabledGroup();
         EditorGUILayout.EndHorizontal();
 
+        // 여기에 나타남
         factScrollPos = EditorGUILayout.BeginScrollView(factScrollPos, EditorStyles.helpBox);
         for (int i = 0; i < FactContainer.Keys.Count; i++)
         {
@@ -75,21 +82,31 @@ public class DyConvWindow : EditorWindow
             factHorizontalStyle.normal.background = (FactContainer.Keys[i] == currentFactName ? Texture2D.grayTexture : Texture2D.blackTexture);
             EditorGUILayout.BeginHorizontal(factHorizontalStyle);
 
+            //누르면 현재 Fact를 선택하고, Dialogue 선택과 Viewer 창을 초기화
             var buttonStyle = new GUIStyle(GUI.skin.label) { alignment = TextAnchor.MiddleLeft, fontSize = 13 };
             if (GUILayout.Button(FactContainer.Keys[i], buttonStyle))
             {
                 currentFactName = FactContainer.Keys[i];
                 isActiveViewer = false;
                 currentDialogueName = null;
+                currentSelectCriteriaIndex = -1;
             }
 
             var textFieldStyle = new GUIStyle(EditorStyles.textField) { alignment = TextAnchor.MiddleRight };
+
+            //변수 값 바꾸는 TextField
             EditorGUI.BeginChangeCheck();
-            FactContainer.FactDictionary[FactContainer.Keys[i]] = EditorGUILayout.DelayedIntField(FactContainer.Values[i], textFieldStyle, GUILayout.MaxWidth(125));
+            if(Application.isPlaying)
+                DyConvManager.FactDictionary[FactContainer.Keys[i]] = EditorGUILayout.DelayedIntField(DyConvManager.FactDictionary[FactContainer.Keys[i]], textFieldStyle, GUILayout.MaxWidth(125));
+            else
+                FactContainer.FactDictionary[FactContainer.Keys[i]] = EditorGUILayout.DelayedIntField(FactContainer.Values[i], textFieldStyle, GUILayout.MaxWidth(125));
             if (EditorGUI.EndChangeCheck())
             {
-                EditorUtility.SetDirty(FactContainer.container);
-                AssetDatabase.SaveAssetIfDirty(FactContainer.container);
+                if(!Application.isPlaying)
+                {
+                    EditorUtility.SetDirty(FactContainer.container);
+                    AssetDatabase.SaveAssetIfDirty(FactContainer.container);
+                }
             }
             EditorGUILayout.EndHorizontal();
         }
@@ -97,6 +114,7 @@ public class DyConvWindow : EditorWindow
         EditorGUILayout.EndVertical();
         #endregion
 
+        //Dialogue 선택창을 나타냄
         #region Dialogue
         EditorGUILayout.BeginVertical(GUI.skin.box, GUILayout.MaxHeight(Screen.height / 2));
 
@@ -104,21 +122,26 @@ public class DyConvWindow : EditorWindow
         EditorGUILayout.LabelField("Dialogues", titleStyle);
         if (GUILayout.Button(EditorGUIUtility.IconContent("d_Toolbar Plus@2x").image, EditorStyles.iconButton))
         {
-            AddNewDialogueWindow.Open();
+            if(!Application.isPlaying)
+                AddNewDialogueWindow.Open();
         }
 
         EditorGUI.BeginDisabledGroup(!isActiveViewer || currentDialogueName == null);
         if (GUILayout.Button(EditorGUIUtility.IconContent("d_Toolbar Minus@2x").image, EditorStyles.iconButton))
         {
-            AssetDatabase.DeleteAsset($"Assets/DyConv/Container/Dialogues/{currentDialogueName}.asset");
-            currentDialogueName = null;
-            isActiveViewer = false;
+            if(!Application.isPlaying)
+            {
+                AssetDatabase.DeleteAsset($"Assets/DyConv/Container/Dialogues/{currentDialogueName}.asset");
+                currentDialogueName = null;
+                isActiveViewer = false;
 
-            AssetDatabase.SaveAssets();
+                AssetDatabase.SaveAssets();
+            }
         }
         EditorGUI.EndDisabledGroup();
         EditorGUILayout.EndHorizontal();
-
+        
+        //여기에 나타남
         dialogueScrollPos = EditorGUILayout.BeginScrollView(dialogueScrollPos, EditorStyles.helpBox);
 
         for (int i = 0; i < Dialogue.allDialogues.Count; i++)
@@ -126,6 +149,7 @@ public class DyConvWindow : EditorWindow
             var buttonStyle = new GUIStyle(EditorStyles.label) { alignment = TextAnchor.MiddleLeft, fontSize = 13 };
             buttonStyle.normal.background = (Dialogue.allDialogues[i].name == currentDialogueName ? Texture2D.grayTexture : Texture2D.blackTexture);
 
+            //누르면 currentDialogueName 변수를 업데이트하고 FactIndex를 초기화시킴
             if (GUILayout.Button($" {Dialogue.allDialogues[i].name}", buttonStyle))
             {
                 currentDialogueName = Dialogue.allDialogues[i].name;
@@ -133,6 +157,8 @@ public class DyConvWindow : EditorWindow
                 currentFactName = null;
 
                 currentNextDialogueIndex = -1;
+                currentSelectCriteriaIndex = -1;
+                EditorGUI.FocusTextInControl(null);
             }
         }
         EditorGUILayout.EndScrollView();
@@ -140,6 +166,7 @@ public class DyConvWindow : EditorWindow
         #endregion
         EditorGUILayout.EndVertical();
 
+        //현재 Dialogue 상태를 나타냄
         #region Viewer
         EditorGUILayout.BeginVertical(GUI.skin.box, GUILayout.MaxWidth(Screen.width * 0.65f), GUILayout.MaxHeight(Screen.height));
         titleStyle = new GUIStyle(GUI.skin.label) { alignment = TextAnchor.MiddleLeft, fontSize = 15, fontStyle = FontStyle.Bold };
@@ -159,6 +186,7 @@ public class DyConvWindow : EditorWindow
 
     void Viewer()
     {
+        //Dialogue 내용
         #region Text Field
         EditorGUILayout.LabelField("Text", new GUIStyle(GUI.skin.label) { alignment = TextAnchor.MiddleLeft, fontSize = 13 });
 
@@ -173,6 +201,7 @@ public class DyConvWindow : EditorWindow
         EditorGUILayout.BeginVertical(GUI.skin.box);
         #endregion
 
+        //다음 Dialogue로 이동할 수 있는 창(모든 Dialogue 경로를 탐색할 경우, Dialogue 끝으로 간주);
         #region Next Dialogues
         EditorGUILayout.BeginHorizontal();
         EditorGUILayout.LabelField("Next Dialogue");
@@ -215,6 +244,7 @@ public class DyConvWindow : EditorWindow
                 shownCriteriaList.Add(new bool());
         }
 
+        //NextDialogue 수만큼 필드 생성
         for (int i = 0; i < Dialogue.GetDialogue(currentDialogueName).nextDialogues.Count; i++)
         {
             int nextDialogueIndex = i;
@@ -225,15 +255,19 @@ public class DyConvWindow : EditorWindow
             var nextDialogueHorizontalStyle = new GUIStyle() { fixedHeight = 25 };
             nextDialogueHorizontalStyle.normal.background = (currentNextDialogueIndex == nextDialogueIndex ? Texture2D.grayTexture : Texture2D.blackTexture);
 
+            //이동하게 될 Dialogue경로와 바꿀수 있는 버튼
             EditorGUILayout.BeginHorizontal(nextDialogueHorizontalStyle);
             shownNextDialogueList[i] = EditorGUILayout.Foldout(shownNextDialogueList[i], nextDialogue.NextDialogueName,
                  new GUIStyle(EditorStyles.foldout) { fixedWidth = 70,fixedHeight = 25 , fontSize = 13 });
 
+            //현재 NextDialogue를 선택
             if (GUILayout.Button("", GUI.skin.label, GUILayout.Height(25)))
             {
                 currentNextDialogueIndex = i;
+                currentSelectCriteriaIndex = -1;
             }
 
+            //NextDialogue의 경로 변경
             if (GUILayout.Button(new GUIContent($" Select", EditorGUIUtility.IconContent("d_RotateTool On").image), new GUIStyle(GUI.skin.button) { alignment = TextAnchor.MiddleCenter }, GUILayout.Width(100)))
             {
                 DialogueSerachPopupWindow.Open((x) =>
@@ -245,12 +279,15 @@ public class DyConvWindow : EditorWindow
             }
             EditorGUILayout.EndHorizontal();
 
+            //Criteria 보여주기
             if (shownNextDialogueList[nextDialogueIndex])
             {
                 EditorGUILayout.BeginVertical(EditorStyles.objectFieldThumb);
 
                 EditorGUILayout.BeginHorizontal();
                 shownCriteriaList[nextDialogueIndex] = EditorGUILayout.Foldout(shownCriteriaList[nextDialogueIndex], "Criteria");
+
+                EditorGUI.BeginDisabledGroup(!shownCriteriaList[i]);
                 if (GUILayout.Button(EditorGUIUtility.IconContent("d_Toolbar Plus@2x").image, EditorStyles.iconButton))
                 {
                     nextDialogue.criteriaKeys.Add("test");
@@ -263,19 +300,34 @@ public class DyConvWindow : EditorWindow
                     shownCriteriaList.Add(new bool());
                 }
 
+                EditorGUI.BeginDisabledGroup(currentSelectCriteriaDialogueIndex != nextDialogueIndex || currentSelectCriteriaIndex == -1);
                 if (GUILayout.Button(EditorGUIUtility.IconContent("d_Toolbar Minus@2x").image, EditorStyles.iconButton))
                 {
+                    nextDialogue.criteriaKeys.RemoveAt(currentSelectCriteriaIndex);
+                    nextDialogue.compareTypes.RemoveAt(currentSelectCriteriaIndex);
+                    nextDialogue.criteriaValues.RemoveAt(currentSelectCriteriaIndex);
 
+                    EditorUtility.SetDirty(Dialogue.GetDialogue(currentDialogueName));
+                    AssetDatabase.SaveAssetIfDirty(Dialogue.GetDialogue(currentDialogueName));
                 }
                 EditorGUILayout.EndHorizontal();
+                EditorGUI.EndDisabledGroup();
+                EditorGUI.EndDisabledGroup();
 
+                //만약 i번째 Criteria가 true라면 보여주기
                 if (shownCriteriaList[i])
                 {
                     criteriaScrollPos = EditorGUILayout.BeginScrollView(criteriaScrollPos, GUILayout.Height(95));
                     for (int j = 0; j < Dialogue.GetDialogue(currentDialogueName).nextDialogues[i].criteriaKeys.Count; j++)
                     {
                         int criteriaIndex = j;
-                        EditorGUILayout.BeginHorizontal(EditorStyles.helpBox);
+
+                        var criteriaStyle = new GUIStyle(EditorStyles.helpBox);
+                        criteriaStyle.normal.background = currentSelectCriteriaDialogueIndex == nextDialogueIndex && currentSelectCriteriaIndex == criteriaIndex ? Texture2D.grayTexture : Texture2D.blackTexture;
+
+                        EditorGUILayout.BeginHorizontal(criteriaStyle);
+
+                        //클릭하면 검사 Fact 교체
                         if (GUILayout.Button(nextDialogue.criteriaKeys[criteriaIndex], new GUIStyle(GUI.skin.button) { alignment = TextAnchor.MiddleLeft }))
                         {
                             FactSearchablePopupWindow.Open((x) =>
@@ -285,11 +337,22 @@ public class DyConvWindow : EditorWindow
                                 AssetDatabase.SaveAssetIfDirty(Dialogue.GetDialogue(currentDialogueName));
                             });
                         }
+
+                        //같다, 다르다, 작다, 크다
                         string[] option = new string[] { "==", "!=", "<", ">" };
                         Dialogue.GetDialogue(currentDialogueName).nextDialogues[i].compareTypes[j] = (CompareType)EditorGUILayout.Popup((int)Dialogue.GetDialogue(currentDialogueName).nextDialogues[i].compareTypes[j], option,
                             new GUIStyle(EditorStyles.popup) { alignment = TextAnchor.MiddleCenter }, GUILayout.MaxWidth(40));
+                        
+                        //검사할 값
                         Dialogue.GetDialogue(currentDialogueName).nextDialogues[i].criteriaValues[j] = EditorGUILayout.IntField(Dialogue.GetDialogue(currentDialogueName).nextDialogues[i].criteriaValues[j],
                             new GUIStyle(GUI.skin.textField) { alignment = TextAnchor.MiddleRight }, GUILayout.MaxWidth(175));
+
+                        //현재 Criteria index 선택
+                        if(GUILayout.Button("☰", EditorStyles.label,GUILayout.MaxWidth(20)))
+                        {
+                            currentSelectCriteriaDialogueIndex = nextDialogueIndex;
+                            currentSelectCriteriaIndex = criteriaIndex;
+                        }
                         EditorGUILayout.EndHorizontal();
                     }
                     EditorGUILayout.EndScrollView();
